+++ examples/advanced/src/gen/mocks/createPetNotFoundFaker.ts
@@ -1,7 +1,7 @@
 import type { PetNotFound } from '../models/ts/PetNotFound.ts'
 import { faker } from '@faker-js/faker'
 
-export function createPetNotFoundFaker(data?: Partial<PetNotFound>): Partial<PetNotFound> {
+export function createPetNotFoundFaker(data?: Partial<PetNotFound>): PetNotFound {
   return {
     ...{ code: faker.number.int(), message: faker.string.alpha() },
     ...(data || {}),