+++ examples/advanced/src/gen/mocks/tag/createTagFaker.ts
@@ -1,7 +1,7 @@
 import type { TagTag } from '../../models/ts/tag/Tag.ts'
 import { faker } from '@faker-js/faker'
 
-export function createTagTagFaker(data?: Partial<TagTag>): Partial<TagTag> {
+export function createTagTagFaker(data?: Partial<TagTag>): TagTag {
   return {
     ...{ id: faker.number.int({ min: 5, max: 7 }), name: faker.string.alpha() },
     ...(data || {}),