+++ examples/react-query/src/gen/hooks/pet/useUploadFileHook.ts
@@ -25,8 +25,8 @@ export async function uploadFileHook(
     url: `/pet/${petId}/uploadImage`,
     params,
     data,
-    headers: { 'Content-Type': 'application/octet-stream', ...requestConfig.headers },
     ...requestConfig,
+    headers: { 'Content-Type': 'application/octet-stream', ...requestConfig.headers },
   })
   return res.data
 }