+++ examples/typescript/src/gen/modelsLiteral.ts
@@ -506,15 +506,22 @@ export type DeletePetHeaderParams = {
   api_key?: string
 }
 
+export type DeletePet200Enum = 'TYPE1' | 'TYPE2' | 'TYPE3'
+
+/**
+ * @description items
+ */
+export type DeletePet200 = DeletePet200Enum[]
+
 /**
  * @description Invalid pet value
  */
 export type DeletePet400 = any
 
-export type DeletePetMutationResponse = any
+export type DeletePetMutationResponse = DeletePet200
 
 export type DeletePetMutation = {
-  Response: any
+  Response: DeletePet200
   PathParams: DeletePetPathParams
   HeaderParams: DeletePetHeaderParams
   Errors: DeletePet400