+++ examples/typescript/src/gen/ts/models/oas.ts
@@ -409,6 +409,20 @@ export const oas = {
           },
         ],
         responses: {
+          '200': {
+            description: 'items',
+            content: {
+              'application/json': {
+                schema: {
+                  type: 'array',
+                  items: {
+                    type: 'string',
+                    enum: ['TYPE1', 'TYPE2', 'TYPE3'],
+                  },
+                },
+              },
+            },
+          },
           '400': {
             description: 'Invalid pet value',
           },